requirements.txt
alembic==1.7.5
amqp==5.0.9
aniso8601==9.0.1
billiard==3.6.4.0
celery==5.2.3
Cerberus==1.3.4
certifi==2021.10.8
charset-normalizer==2.0.10
clean-architecture==0.0.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contextlib2==21.6.0
distlib==0.3.4
filelock==3.4.2
Flask==2.0.2
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
kombu==5.2.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
packaging==21.3
passlib==1.7.4
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
py==1.11.0
pybit==1.3.5
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
schema==0.7.5
six==1.16.0
SQLAlchemy==1.4.30
sqlalchemy-filters==0.12.0
toml==0.10.2
tox==3.24.5
urllib3==1.26.8
vine==5.0.0
virtualenv==20.13.0
wcwidth==0.2.5
websocket-client==1.2.3
websockets==10.1
Werkzeug==2.0.2
